Selenium Code
Got Jest running and my finished selenium code. Passed 87% of tests. Fails for creating an answer key with the correct number of questions. Currently if you edit the answer key there are always 100 questions. A newly created exam also does not stay on the homepage after navigating through the site.