Fixed Arithmetic Exception
In the case that there is no variance between exam scores or variance between question scores, the variance would be zero. This would cause a division by zero. I handled this by making it so that if it divides by zero, it will instead return a 1, indicating that the exam tested what you wanted it to test, seeing as how everyone got a 100 or the same score.